Move .as_dict() from Build() to LazyDatabaseMixin()
