+++ test/test_build.py
@@ -398,24 +398,3 @@ class TestBuildApiCreate(object):
         }
 
         api.extract_json.assert_called_once_with(post)
-
-
-class TestBuildAsDict(object):
-    def test_without_id_key(self, build):
-        ret = build.as_dict()
-        assert 'id' not in ret
-
-    def test_with_id_key(self, build):
-        build.id = True  # Can't use a mock because it will have a .raw
-        ret = build.as_dict()
-        assert ret['id'] is build.id
-
-    @mock.patch('piper.utils.now')
-    def test_timestamp_is_added(self, now, build):
-        ret = build.as_dict()
-        assert ret['updated'] is now.return_value
-
-    def test_raw_value(self, build):
-        build.config = Mock()
-        ret = build.as_dict()
-        assert ret['config'] is build.config.raw
